old/requirements.txt
Django==2.1.3
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-extensions==2.1.4
django-js-asset==1.1.0
django-rest-framework==0.1.0
django-tinymce==2.7.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
six==1.12.0
